Fix compiler warning C4267 re-introduced by r7159.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7193 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/DownloadTransferHTTP.cpp
@@ -1292,7 +1292,7 @@ BOOL CDownloadTransferHTTP::ReadContent()
 	{
 		m_pSource->SetValid();
 
-		DWORD nLength	= (DWORD)min( pInput->m_nLength, m_nLength - m_nPosition );
+		size_t nLength	= min( pInput->m_nLength, m_nLength - m_nPosition );
 		BOOL bSubmit	= FALSE;
 
 		if ( m_bChunked )
